src/serving/requirements.txt
flask==3.0.0
xgboost==2.0.0
pandas==2.1.0
scikit-learn==1.3.0
google-cloud-storage==2.13.0
gunicorn==21.2.0
